Add key_hash for smaller index
This adds a key_hash column to the solid_cache_entries table. This
column is a 64-bit integer that is a truncated SHA256 hash of the key.
This allows us to use a smaller index than the 1024 byte key column.

The byte_size column is also added to the table. This is the size of
the value in bytes. This is for a separate feature but is included in
this commit to piggy back on the migrations.

The migrations themselves are split into three steps. This allows
existing tables to be migrated without downtime.

1. The first migration add nullable columns. After a restart the new
   columns will populated for new records. Then the tables can be
   backfilled or truncated.

2. The second migration adds indexes and null constraints. After a
   restart the key_hash index will be used for lookups.

3. Then the third migration removes the old key index.

The code will work at any stage of the migration process. It checks for
the existance of the key_hash column and index and populates it when it
exists and queries on it when it is indexed.

For later versions of the gem we'll assume that the key_hash column
exists and remove these checks. We'll also squash the migrations into
a single one.